Split HLOOKUP(), VLOOKUP() and LOOKUP() into individual classes
