+++ src/PhpSpreadsheet/Calculation/LookupRef/Lookup.php
@@ -2,120 +2,11 @@
 
 namespace PhpOffice\PhpSpreadsheet\Calculation\LookupRef;
 
-use PhpOffice\PhpSpreadsheet\Calculation\Exception;
 use PhpOffice\PhpSpreadsheet\Calculation\Functions;
 use PhpOffice\PhpSpreadsheet\Calculation\LookupRef;
-use PhpOffice\PhpSpreadsheet\Shared\StringHelper;
 
 class Lookup
 {
-    private static function validateIndexLookup($lookup_array, $index_number)
-    {
-        // index_number must be a number greater than or equal to 1
-        if (!is_numeric($index_number) || $index_number < 1) {
-            throw new Exception(Functions::VALUE());
-        }
-
-        // index_number must be less than or equal to the number of columns in lookup_array
-        if ((!is_array($lookup_array)) || (empty($lookup_array))) {
-            throw new Exception(Functions::REF());
-        }
-
-        return (int) $index_number;
-    }
-
-    /**
-     * VLOOKUP
-     * The VLOOKUP function searches for value in the left-most column of lookup_array and returns the value
-     *     in the same row based on the index_number.
-     *
-     * @param mixed $lookupValue The value that you want to match in lookup_array
-     * @param mixed $lookupArray The range of cells being searched
-     * @param mixed $indexNumber The column number in table_array from which the matching value must be returned.
-     *                                The first column is 1.
-     * @param mixed $notExactMatch determines if you are looking for an exact match based on lookup_value
-     *
-     * @return mixed The value of the found cell
-     */
-    public static function VLOOKUP($lookupValue, $lookupArray, $indexNumber, $notExactMatch = true)
-    {
-        $lookupValue = Functions::flattenSingleValue($lookupValue);
-        $indexNumber = Functions::flattenSingleValue($indexNumber);
-        $notExactMatch = Functions::flattenSingleValue($notExactMatch);
-
-        try {
-            $indexNumber = self::validateIndexLookup($lookupArray, $indexNumber);
-        } catch (Exception $e) {
-            return $e->getMessage();
-        }
-
-        $f = array_keys($lookupArray);
-        $firstRow = array_pop($f);
-        if ((!is_array($lookupArray[$firstRow])) || ($indexNumber > count($lookupArray[$firstRow]))) {
-            return Functions::REF();
-        }
-        $columnKeys = array_keys($lookupArray[$firstRow]);
-        $returnColumn = $columnKeys[--$indexNumber];
-        $firstColumn = array_shift($columnKeys);
-
-        if (!$notExactMatch) {
-            uasort($lookupArray, ['self', 'vlookupSort']);
-        }
-
-        $rowNumber = self::vLookupSearch($lookupValue, $lookupArray, $firstColumn, $notExactMatch);
-
-        if ($rowNumber !== null) {
-            // return the appropriate value
-            return $lookupArray[$rowNumber][$returnColumn];
-        }
-
-        return Functions::NA();
-    }
-
-    /**
-     * HLOOKUP
-     * The HLOOKUP function searches for value in the top-most row of lookup_array and returns the value
-     *     in the same column based on the index_number.
-     *
-     * @param mixed $lookupValue The value that you want to match in lookup_array
-     * @param mixed $lookupArray The range of cells being searched
-     * @param mixed $indexNumber The row number in table_array from which the matching value must be returned.
-     *                                The first row is 1.
-     * @param mixed $notExactMatch determines if you are looking for an exact match based on lookup_value
-     *
-     * @return mixed The value of the found cell
-     */
-    public static function HLOOKUP($lookupValue, $lookupArray, $indexNumber, $notExactMatch = true)
-    {
-        $lookupValue = Functions::flattenSingleValue($lookupValue);
-        $indexNumber = Functions::flattenSingleValue($indexNumber);
-        $notExactMatch = Functions::flattenSingleValue($notExactMatch);
-
-        try {
-            $indexNumber = self::validateIndexLookup($lookupArray, $indexNumber);
-        } catch (Exception $e) {
-            return $e->getMessage();
-        }
-
-        $f = array_keys($lookupArray);
-        $firstRow = reset($f);
-        if ((!is_array($lookupArray[$firstRow])) || ($indexNumber > count($lookupArray))) {
-            return Functions::REF();
-        }
-
-        $firstkey = $f[0] - 1;
-        $returnColumn = $firstkey + $indexNumber;
-        $firstColumn = array_shift($f);
-        $rowNumber = self::hLookupSearch($lookupValue, $lookupArray, $firstColumn, $notExactMatch);
-
-        if ($rowNumber !== null) {
-            //  otherwise return the appropriate value
-            return $lookupArray[$returnColumn][$rowNumber];
-        }
-
-        return Functions::NA();
-    }
-
     /**
      * LOOKUP
      * The LOOKUP function searches for value either from a one-row or one-column range or from an array.
@@ -126,7 +17,7 @@ class Lookup
      *
      * @return mixed The value of the found cell
      */
-    public static function LOOKUP($lookupValue, $lookupVector, $resultVector = null)
+    public static function lookup($lookupValue, $lookupVector, $resultVector = null)
     {
         $lookupValue = Functions::flattenSingleValue($lookupValue);
 
@@ -184,106 +75,6 @@ class Lookup
             unset($value);
         }
 
-        return self::VLOOKUP($lookupValue, $lookupVector, 2);
-    }
-
-    private static function vlookupSort($a, $b)
-    {
-        reset($a);
-        $firstColumn = key($a);
-        $aLower = StringHelper::strToLower($a[$firstColumn]);
-        $bLower = StringHelper::strToLower($b[$firstColumn]);
-        if ($aLower == $bLower) {
-            return 0;
-        }
-
-        return ($aLower < $bLower) ? -1 : 1;
-    }
-
-    private static function vLookupSearch($lookupValue, $lookupArray, $column, $notExactMatch)
-    {
-        $lookupLower = StringHelper::strToLower($lookupValue);
-
-        $rowNumber = null;
-        foreach ($lookupArray as $rowKey => $rowData) {
-            $bothNumeric = is_numeric($lookupValue) && is_numeric($rowData[$column]);
-            $bothNotNumeric = !is_numeric($lookupValue) && !is_numeric($rowData[$column]);
-            $cellDataLower = StringHelper::strToLower($rowData[$column]);
-
-            // break if we have passed possible keys
-            if ($notExactMatch &&
-                ($bothNumeric && ($rowData[$column] > $lookupValue)) ||
-                ($bothNotNumeric && ($cellDataLower > $lookupLower))
-            ) {
-                break;
-            }
-
-            $rowNumber = self::checkMatch(
-                $bothNumeric,
-                $bothNotNumeric,
-                $notExactMatch,
-                $rowKey,
-                $cellDataLower,
-                $lookupLower,
-                $rowNumber
-            );
-        }
-
-        return $rowNumber;
-    }
-
-    private static function hLookupSearch($lookupValue, $lookupArray, $column, $notExactMatch)
-    {
-        $lookupLower = StringHelper::strToLower($lookupValue);
-
-        $rowNumber = null;
-        foreach ($lookupArray[$column] as $rowKey => $rowData) {
-            // break if we have passed possible keys
-            $bothNumeric = is_numeric($lookupValue) && is_numeric($rowData);
-            $bothNotNumeric = !is_numeric($lookupValue) && !is_numeric($rowData);
-            $cellDataLower = StringHelper::strToLower($rowData);
-
-            if ($notExactMatch &&
-                (($bothNumeric && $rowData > $lookupValue) || ($bothNotNumeric && $cellDataLower > $lookupLower))
-            ) {
-                break;
-            }
-
-            $rowNumber = self::checkMatch(
-                $bothNumeric,
-                $bothNotNumeric,
-                $notExactMatch,
-                $rowKey,
-                $cellDataLower,
-                $lookupLower,
-                $rowNumber
-            );
-        }
-
-        return $rowNumber;
-    }
-
-    private static function checkMatch(
-        bool $bothNumeric,
-        bool $bothNotNumeric,
-        $notExactMatch,
-        int $rowKey,
-        string $cellDataLower,
-        string $lookupLower,
-        ?int $rowNumber
-    ): ?int {
-        // remember the last key, but only if datatypes match
-        if ($bothNumeric || $bothNotNumeric) {
-            // Spreadsheets software returns first exact match,
-            // we have sorted and we might have broken key orders
-            // we want the first one (by its initial index)
-            if ($notExactMatch) {
-                $rowNumber = $rowKey;
-            } elseif (($cellDataLower == $lookupLower) && (($rowNumber === null) || ($rowKey < $rowNumber))) {
-                $rowNumber = $rowKey;
-            }
-        }
-
-        return $rowNumber;
+        return VLookup::lookup($lookupValue, $lookupVector, 2);
     }
 }
